Add files via upload
Extensions and Improvements

There are several ways in which the application and API could be extended and improved:

User authentication: Currently, anyone can submit a response to the survey. Adding user authentication would allow only authorized users to submit responses.

Validation: Adding validation to the form fields would ensure that users enter valid data. For example, the email address field could be validated to ensure that it is in the correct format.

Data visualization: Adding charts and graphs to the survey results would make it easier to visualize the data and identify trends.

Deployment

To deploy the application and API, I would recommend using a cloud hosting service such as Heroku or AWS. Additionally, I would recommend setting up a CI/CD pipeline to automate the deployment process and ensure that the application is always up-to-date and running smoothly.